+++ src/main/java/shadows/apotheosis/core/attributeslib/AttributesLib.java
@@ -16,6 +16,7 @@ import net.minecraftforge.registries.RegistryObject;
 import shadows.apotheosis.core.attributeslib.api.ALAttributes;
 import shadows.apotheosis.core.attributeslib.client.AttributesLibClient;
 import shadows.apotheosis.core.attributeslib.impl.AttributeEvents;
+import shadows.apotheosis.core.attributeslib.impl.PercentBasedAttribute;
 import shadows.placebo.util.RegistryEvent.Register;
 
 public class AttributesLib {
@@ -33,25 +34,29 @@ public class AttributesLib {
 	public void attribs(Register<Attribute> e) {
 		//Formatter::off
 		e.getRegistry().registerAll(
-			new RangedAttribute("apotheosis:draw_speed", 1.0D, 1.0D, 4.0D).setSyncable(true), "draw_speed",
-			new RangedAttribute("apotheosis:crit_chance", 1.0D, 1.0D, 1024.0D).setSyncable(true), "crit_chance",
-			new RangedAttribute("apotheosis:crit_damage", 1.5D, 1.0D, 1024.0D).setSyncable(true), "crit_damage",
+			new PercentBasedAttribute("apotheosis:draw_speed", 1.0D, 1.0D, 4.0D).setSyncable(true), "draw_speed",
+			new PercentBasedAttribute("apotheosis:crit_chance", 0.05D, 0.0D, 1024.0D).setSyncable(true), "crit_chance",
+			new PercentBasedAttribute("apotheosis:crit_damage", 1.5D, 1.0D, 1024.0D).setSyncable(true), "crit_damage",
 			new RangedAttribute("apotheosis:cold_damage", 0.0D, 0.0D, 1024.0D).setSyncable(true), "cold_damage",
 			new RangedAttribute("apotheosis:fire_damage", 0.0D, 0.0D, 1024.0D).setSyncable(true), "fire_damage",
-			new RangedAttribute("apotheosis:life_steal", 1.0D, 1.0D, 1024.0D).setSyncable(true), "life_steal",
-			new RangedAttribute("apotheosis:piercing", 1.0D, 1.0D, 2.0D).setSyncable(true), "piercing",
-			new RangedAttribute("apotheosis:current_hp_damage", 1.0D, 1.0D, 2.0D).setSyncable(true), "current_hp_damage",
-			new RangedAttribute("apotheosis:overheal", 1.0D, 0.0D, 1024.0D).setSyncable(true), "overheal",
+			new PercentBasedAttribute("apotheosis:life_steal", 0.0D, 0.0D, 1024.0D).setSyncable(true), "life_steal",
+			new PercentBasedAttribute("apotheosis:current_hp_damage", 0.0D, 0.0D, 2.0D).setSyncable(true), "current_hp_damage",
+			new PercentBasedAttribute("apotheosis:overheal", 0.0D, 0.0D, 1024.0D).setSyncable(true), "overheal",
 			new RangedAttribute("apotheosis:ghost_health", 0.0D, 0.0D, 1024.0D).setSyncable(true), "ghost_health",
-			new RangedAttribute("apotheosis:mining_speed", 1.0D, 0.0D, 1024.0D).setSyncable(true), "mining_speed",
-			new RangedAttribute("apotheosis:arrow_damage", 1.0D, 0.0D, 1024.0D).setSyncable(true), "arrow_damage",
-			new RangedAttribute("apotheosis:arrow_velocity", 1.0D, 0.0D, 1024.0D).setSyncable(true), "arrow_velocity",
-			new RangedAttribute("apotheosis:experience_gained", 1.0D, 0.0D, 1024.0D).setSyncable(true), "experience_gained",
-			new RangedAttribute("apotheosis:healing_received", 1.0D, 0.0D, 1024.0D).setSyncable(true), "healing_received"
+			new PercentBasedAttribute("apotheosis:mining_speed", 1.0D, 0.0D, 1024.0D).setSyncable(true), "mining_speed",
+			new PercentBasedAttribute("apotheosis:arrow_damage", 1.0D, 0.0D, 1024.0D).setSyncable(true), "arrow_damage",
+			new PercentBasedAttribute("apotheosis:arrow_velocity", 1.0D, 0.0D, 1024.0D).setSyncable(true), "arrow_velocity",
+			new PercentBasedAttribute("apotheosis:experience_gained", 1.0D, 0.0D, 1024.0D).setSyncable(true), "experience_gained",
+			new PercentBasedAttribute("apotheosis:healing_received", 1.0D, 0.0D, 1024.0D).setSyncable(true), "healing_received",
+			new RangedAttribute("apotheosis:armor_piercing", 0.0D, 0.0D, 1024.0D).setSyncable(true), "armor_pierce",
+			new PercentBasedAttribute("apotheosis:armor_shred", 0.0D, 0.0D, 1.0D).setSyncable(true), "armor_shred",
+			new RangedAttribute("apotheosis:prot_piercing", 0.0D, 0.0D, 1024.0D).setSyncable(true), "prot_pierce",
+			new PercentBasedAttribute("apotheosis:prot_shred", 0.0D, 0.0D, 1.0D).setSyncable(true), "prot_shred"
 		);
 		//Formatter::on
 	}
 
+	// TODO - Update impls to reflect new default values.
 	@SubscribeEvent
 	public void applyAttribs(EntityAttributeModificationEvent e) {
 		e.getTypes().forEach(type -> {
@@ -63,7 +68,6 @@ public class AttributesLib {
 					ALAttributes.COLD_DAMAGE,
 					ALAttributes.FIRE_DAMAGE,
 					ALAttributes.LIFE_STEAL,
-					ALAttributes.PIERCING,
 					ALAttributes.CURRENT_HP_DAMAGE,
 					ALAttributes.OVERHEAL,
 					ALAttributes.GHOST_HEALTH,
@@ -71,7 +75,12 @@ public class AttributesLib {
 					ALAttributes.ARROW_DAMAGE,
 					ALAttributes.ARROW_VELOCITY,
 					ALAttributes.EXPERIENCE_GAINED,
-					ALAttributes.HEALING_RECEIVED);
+					ALAttributes.HEALING_RECEIVED,
+					ALAttributes.ARMOR_PIERCE,
+					ALAttributes.ARMOR_SHRED,
+					ALAttributes.PROT_PIERCE,
+					ALAttributes.PROT_SHRED
+					);
 			//Formatter::on
 		});
 	}