+++ src/main/java/shadows/apotheosis/core/attributeslib/api/ALAttributes.java
@@ -11,18 +11,19 @@ public class ALAttributes {
 	private static final RegObjHelper R = new RegObjHelper(Apotheosis.MODID);
 
 	/**
-	 * Bonus to how fast a ranged weapon is charged. Base Value = (1.0) = 100%
+	 * How fast a ranged weapon is charged. Base Value = (1.0) = 100% default draw speed.
 	 */
 	public static final RegistryObject<Attribute> DRAW_SPEED = R.attribute("draw_speed");
 
 	/**
-	 * Chance that a non-jump-attack will critically strike.  Base value = (1.0) = 0%
+	 * Chance that any attack will critically strike.  Base value = (0.05) = 5% chance to critically strike.<br>
+	 * Not related to vanilla (jump) critical strikes.
 	 */
 	public static final RegistryObject<Attribute> CRIT_CHANCE = R.attribute("crit_chance");
 
 	/**
-	 * Amount of damage caused by critical strikes. Base value = (1.0) = 100%
-	 * Not related to vanilla critical strikes.
+	 * Amount of damage caused by critical strikes. Base value = (1.5) = 150% normal damage dealt.<br>
+	 * Also impacts vanilla (jump) critical strikes.
 	 */
 	public static final RegistryObject<Attribute> CRIT_DAMAGE = R.attribute("crit_damage");
 
@@ -37,22 +38,17 @@ public class ALAttributes {
 	public static final RegistryObject<Attribute> FIRE_DAMAGE = R.attribute("fire_damage");
 
 	/**
-	 * Percent of physical damage converted to health. Base value = (1.0) = 0%
+	 * Percent of physical damage converted to health. Base value = (0.0) = 0%
 	 */
 	public static final RegistryObject<Attribute> LIFE_STEAL = R.attribute("life_steal");
 
 	/**
-	 * Percent of physical damage that bypasses armor. Base value = (1.0) = 0%
-	 */
-	public static final RegistryObject<Attribute> PIERCING = R.attribute("piercing");
-
-	/**
-	 * Bonus physical damage dealt equal to enemy's current health. Base value = (1.0) = 0%
+	 * Bonus physical damage dealt equal to enemy's current health. Base value = (0.0) = 0%
 	 */
 	public static final RegistryObject<Attribute> CURRENT_HP_DAMAGE = R.attribute("current_hp_damage");
 
 	/**
-	 * Percent of physical damage converted to absorption hearts. Base value = (1.0) = 0%
+	 * Percent of physical damage converted to absorption hearts. Base value = (0.0) = 0%
 	 */
 	public static final RegistryObject<Attribute> OVERHEAL = R.attribute("overheal");
 
@@ -85,5 +81,25 @@ public class ALAttributes {
 	 * Adjusts all healing received. Base value = (1.0) = 100% xp gained.
 	 */
 	public static final RegistryObject<Attribute> HEALING_RECEIVED = R.attribute("healing_received");
+
+	/**
+	 * Flat armor penetration. Base value = (0.0) = 0 armor reduced during damage calculations.
+	 */
+	public static final RegistryObject<Attribute> ARMOR_PIERCE = R.attribute("armor_pierce");
+
+	/**
+	 * Percentage armor reduction. Base value = (0.0) = 0% of armor reduced during damage calculations.
+	 */
+	public static final RegistryObject<Attribute> ARMOR_SHRED = R.attribute("armor_shred");
+
+	/**
+	 * Flat protection penetration. Base value = (0.0) = 0 protection points bypassed during damage calculations.
+	 */
+	public static final RegistryObject<Attribute> PROT_PIERCE = R.attribute("prot_pierce");
+
+	/**
+	 * Percentage protection reduction. Base value = (0.0) = 0% of protection points bypassed during damage calculations.
+	 */
+	public static final RegistryObject<Attribute> PROT_SHRED = R.attribute("prot_shred");
 
 }