+++ src/main/java/shadows/apotheosis/mixin/LivingEntityMixin.java
@@ -5,7 +5,10 @@ import javax.annotation.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.Surrogate;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.damagesource.DamageSource;
@@ -16,6 +19,7 @@ import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
 import shadows.apotheosis.Apoth;
+import shadows.apotheosis.core.attributeslib.asm.ALCombatRules;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin extends Entity {
@@ -33,7 +37,7 @@ public abstract class LivingEntityMixin extends Entity {
 	 * @param damage The initial damage amount
 	 */
 	@Redirect(at = @At(value = "INVOKE", target = "Ljava/lang/Math;max(FF)F"), method = "getDamageAfterMagicAbsorb(Lnet/minecraft/world/damagesource/DamageSource;F)F")
-	public float apoth_sundering_max(float value, float max, DamageSource source, float damage) {
+	public float apoth_sunderingApplyEffect(float value, float max, DamageSource source, float damage) {
 		if (Apoth.Effects.SUNDERING.isPresent() && this.hasEffect(Apoth.Effects.SUNDERING.get()) && source != DamageSource.OUT_OF_WORLD) {
 			int level = this.getEffect(Apoth.Effects.SUNDERING.get()).getAmplifier() + 1;
 			value += damage * level * 0.2F;
@@ -46,7 +50,7 @@ public abstract class LivingEntityMixin extends Entity {
 	 * @reason Used to enter an if-condition so the above mixin always triggers.
 	 */
 	@Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/LivingEntity;hasEffect(Lnet/minecraft/world/effect/MobEffect;)Z"), method = "getDamageAfterMagicAbsorb(Lnet/minecraft/world/damagesource/DamageSource;F)F")
-	public boolean apoth_sundering_hasEffect(LivingEntity ths, MobEffect effect) {
+	public boolean apoth_sunderingHasEffect(LivingEntity ths, MobEffect effect) {
 		return true;
 	}
 
@@ -55,7 +59,7 @@ public abstract class LivingEntityMixin extends Entity {
 	 * @reason Used to prevent an NPE since we're faking true on hasEffect
 	 */
 	@Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/effect/MobEffectInstance;getAmplifier()I"), method = "getDamageAfterMagicAbsorb(Lnet/minecraft/world/damagesource/DamageSource;F)F")
-	public int apoth_sundering_getAmplifier(@Nullable MobEffectInstance inst) {
+	public int apoth_sunderingGetAmplifier(@Nullable MobEffectInstance inst) {
 		return inst == null ? -1 : inst.getAmplifier();
 	}
 
@@ -65,13 +69,23 @@ public abstract class LivingEntityMixin extends Entity {
 	@Shadow
 	public abstract MobEffectInstance getEffect(MobEffect ef);
 
-	@Override
-	public int getTeamColor() {
-		if (super.getTeamColor() == 16777215) {
+	@Surrogate
+	@Inject(at = @At(value = "RETURN"), method = "getTeamColor()I")
+	public void apoth_applyBossGlowingColor(CallbackInfoReturnable<Integer> cir) {
+		if (cir.getReturnValueI() == 16777215) {
 			Component name = this.getCustomName();
-			if (name != null && name.getStyle().getColor() != null) return name.getStyle().getColor().getValue();
+			if (name != null && name.getStyle().getColor() != null) cir.setReturnValue(name.getStyle().getColor().getValue());
 		}
-		return super.getTeamColor();
+	}
+
+	@Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/damagesource/CombatRules;getDamageAfterAbsorb(FFF)F"), method = "getDamageAfterArmorAbsorb(Lnet/minecraft/world/damagesource/DamageSource;F)F", require = 1)
+	public float apoth_applyArmorPen(float amount, float armor, float toughness, DamageSource src, float amt2) {
+		return ALCombatRules.getDamageAfterArmor((LivingEntity) (Object) this, src, amount, armor, toughness);
+	}
+
+	@Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/damagesource/CombatRules;getDamageAfterMagicAbsorb(FF)F"), method = "getDamageAfterMagicAbsorb(Lnet/minecraft/world/damagesource/DamageSource;F)F", require = 1)
+	public float apoth_applyProtPen(float amount, float protPoints, DamageSource src, float amt2) {
+		return ALCombatRules.getDamageAfterProtection((LivingEntity) (Object) this, src, amount, protPoints);
 	}
 
 }